fix: 10 correctness bugs across engine core
- fsync/fdatasync use RING_SUBMIT_META_OP to stop polluting AIMD P99
- poll() non-EINTR errors in aura_wait now return -1 instead of 0
- fix UAF in tls_destructor: always clear pthread key when freeing cache
- fix buffer_pool_free race with destroy via cleanup_mutex serialization
- ring_destroy: close io_uring before freeing requests to prevent stale CQEs
- aura_run drain timeout logs warning when abandoning in-flight I/O
- aura_histogram_percentile p100 clamped to max_tracked_us
- adaptive_hist_p99 accumulator widened to uint64_t to prevent overflow
- ring_wait retry loop uses memory_order_acquire for pending_count
- tick_stats_t.sample_count changed from int to uint32_t

+++ engine/src/adaptive_engine.c
@@ -50,7 +50,7 @@ static double window_min(const double *window, int count) {
  * Used to pass computed values from tick_swap_and_compute_stats() to state handlers.
  */
 typedef struct {
-    int sample_count;
+    uint32_t sample_count;
     double p99_ms;
     double throughput_bps;
     double sqe_ratio;
@@ -219,8 +219,9 @@ double adaptive_hist_p99(adaptive_histogram_t *hist) {
         if (target == 0) target = 1;
     }
 
-    /* Scan from high to low, counting down */
-    uint32_t count = atomic_load_explicit(&hist->overflow, memory_order_relaxed);
+    /* Scan from high to low, counting down.
+     * Use uint64_t accumulator to prevent overflow at extreme IOPS. */
+    uint64_t count = atomic_load_explicit(&hist->overflow, memory_order_relaxed);
     if (count >= target) {
         /* P99 is in overflow bucket (> 10ms) */
         return (double)LATENCY_MAX_US / 1000.0;
